Add Entity Component System and editor panels - Entity Component System (ECS) 	- Add Components 	- Add Entity, Component and System Managers 	- Add Registry to control all Managers 	Notes: 	- Entity type is uint32_t 	- Component types is uint16_t 	- ECS includes Systems which keep track of which entities 	  have which components 	- Each entity has a Signiture which keeps track of what 	  components it has 	- Signitures are bitsets - Add Editor Panels in Editor 	- Add SceneHierarchy panel 	- Add Properties panel 	- Add Scene to editor to contain entities - Move Core files into own Core folder - Add RendererCommand class for common renderering functions (e.g.   renderClear, renderClearColor) - Add glTF file support for glTF models

+++ Chonps/src/Graphics/IndexBuffer.h
@@ -7,7 +7,7 @@ namespace Chonps
 	{
 	public:
 		// IndexBuffer constructor for static arrays
-		IndexBuffer(uint32_t* indices, size_t size) {}
+		IndexBuffer(uint32_t* indices, uint32_t size) {}
 		// IndexBuffer constructor for dynamic arrays
 		IndexBuffer(std::vector<uint32_t>& indices) {}
 
@@ -20,7 +20,7 @@ namespace Chonps
 	};
 	typedef IndexBuffer IBO;
 
-	IndexBuffer* createIndexBuffer(uint32_t* indices, size_t size);
+	IndexBuffer* createIndexBuffer(uint32_t* indices, uint32_t size);
 	IndexBuffer* createIndexBuffer(std::vector<uint32_t>& indices);
 }
